Cleaned up files in Documents. Added description to README. Fixed a few frontend integration bugs.
