+++ market_magician_dev/frontend/src/App.css
@@ -9,19 +9,81 @@ body {
 }
 
 .header {
-  background-color: #333;
+  background-color: #2F4F4F;
   color: white;
-  padding: 20px;
+  padding: 15px 30px;
   display: flex;
   justify-content: space-between;
   align-items: center;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
 }
 
-.nav a {
-  margin-left: 20px;
+.header .navbar-brand {
+  font-size: 24px;
+  font-weight: bold;
   color: white;
   text-decoration: none;
 }
+
+.nav {
+  display: flex;
+  gap: 20px;
+}
+
+.nav a {
+  color: white;
+  text-decoration: none;
+  font-size: 16px;
+  transition: color 0.3s ease;
+}
+
+.nav a:hover {
+  color: #ccc;
+}
+
+
+.navbar-toggler {
+  display: none;
+  background-color: white;
+  border: none;
+  color: #007bff;
+  font-size: 24px;
+  cursor: pointer;
+}
+
+.nav-link {
+  color: white;
+  text-decoration: none;
+  font-size: 16px;
+  margin-left: 20px;
+}
+
+
+@media (max-width: 768px) {
+  .header {
+    flex-direction: column;
+    align-items: flex-start;
+  }
+
+  .nav {
+    display: none;
+    flex-direction: column;
+    gap: 10px;
+    width: 100%;
+    margin-top: 10px;
+  }
+
+  .nav.show {
+    display: flex;
+  }
+
+  .navbar-toggler {
+    display: block;
+    align-self: flex-end;
+    margin-bottom: 10px;
+  }
+}
+
 
 .main-section {
   padding: 50px;
@@ -45,7 +107,6 @@ body {
   margin-bottom: 20px;
 }
 
-
 .specific-stock-info {
   background-color: #fff;
   padding: 40px;
@@ -61,6 +122,50 @@ body {
   padding: 20px;
   width: 22%;
   border: 1px solid #ddd;
+}
+
+.learn-more-btn {
+  background-color: #007BFF;
+  color: white;
+  border: none;
+  padding: 12px 24px;
+  font-size: 16px;
+  border-radius: 5px;
+  transition: background-color 0.3s ease;
+}
+
+.learn-more-btn:hover {
+  background-color: #0056b3;
+}
+
+.login-form {
+  max-width: 400px;
+  margin: 20px auto;
+  padding: 20px;
+  border: 1px solid #ddd;
+  border-radius: 8px;
+  background-color: #f9f9f9;
+}
+
+.login-form h2 {
+  margin-bottom: 20px;
+}
+
+.form-group {
+  margin-bottom: 15px;
+}
+
+.btn-primary {
+  background-color: #007bff;
+  color: white;
+  border: none;
+  padding: 10px 20px;
+  cursor: pointer;
+  border-radius: 5px;
+}
+
+.btn-primary:hover {
+  background-color: #0056b3;
 }
 
 .footer {
