+++ market_magician_dev/frontend/src/App.js
@@ -1,13 +1,15 @@
-import React from "react";
-import { BrowserRouter, Routes, Route, Link } from "react-router-dom";
-import Header from "./components/Header";
-//import StockInfo from "./components/StockInfo";
-import Footer from "./components/Footer";
-import BuyStocks from "./pages/BuyStocks";
-import Home from "./components/Home";
-import Login from "./components/Login";
-import SignUp from "./components/SignUp";
-import Send_ticker from './pages/Send_ticker';
+import React from 'react';
+import { Routes, Route } from 'react-router-dom';
+
+import Header from './components/Header';
+import Footer from './components/Footer';
+
+import Home from './pages/Home';
+import Login from './pages/Login';
+import SignUp from './pages/SignUp';
+import BuyStocks from './pages/BuyStocks';
+import SearchTicker from './pages/SearchTicker';
+
 import './App.css';
 
 
@@ -18,10 +20,11 @@ function App() {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/buyStocks" element={<BuyStocks />} /> {/* Route for BuyStocks */}
-        <Route path="/components/Login" element={<Login />} />
-        <Route path="/components/SignUp" element={<SignUp />} />
-        <Route path="/sendTicker" element={<Send_ticker />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/signup" element={<SignUp />} />
+        <Route path="/searchTicker" element={<SearchTicker />} />
       </Routes>
+      <Footer />
     </div>
   );
 }
