+++ market_magician_dev/frontend/src/pages/Home.js
