+++ market_magician_dev/predictions/views.py
@@ -24,7 +24,7 @@ from sklearn.metrics import mean_absolute_error, mean_squared_error
 
 # import pandas_datareader as pdr
 import matplotlib
-matplotlib.use('Agg')
+# matplotlib.use('Agg')
 import matplotlib.pyplot as plt
 from tensorflow import keras
 from tensorflow.keras import layers
